Create working add/delete test with mock dom
