+++ src/modules/app.js
@@ -50,3 +50,5 @@ clearButton.addEventListener('click', (e) => {
     listItems[i].setAttribute('id', i + 1);
   }
 });
+
+module.exports = addTask;
